Simplify the structure
Could we perhaps make the case that this should simply leverage Laravel's built-in slug method directly?
