+++ src/SlugSanitizers/BaseSlugSanitizer.php
@@ -6,11 +6,17 @@ use Lukeraymonddowning\SelfHealingUrls\Contracts\SlugSanitizer;
 
 class BaseSlugSanitizer implements SlugSanitizer
 {
+    public function __construct(
+        private readonly string $separator = '-',
+        private readonly array $dictionary = ['@' => ''],
+    ) {
+    }
+
     public function sanitize(string $slug): string
     {
-        return str($slug)
-            ->lower()
-            ->ascii()
-            ->replaceMatches('/[^a-z0-9-_ ]/', '');
+        return str($slug)->slug(
+            separator: $this->separator,
+            dictionary: $this->dictionary,
+        );
     }
 }
