--- tests/Unit/SlugSanitizers/KebabSlugSanitizerTest.php
@@ -1,24 +1,0 @@
-<?php
-
-use Lukeraymonddowning\SelfHealingUrls\SlugSanitizers\StringHelperSlugSanitizer;
-
-it('replaces spaces with hyphens', function () {
-    $sanitizer = new StringHelperSlugSanitizer();
-    $result = $sanitizer->sanitize('what are you doing?');
-
-    expect($result)->toBe('what-are-you-doing');
-});
-
-it('removes multiple hyphens', function () {
-    $sanitizer = new StringHelperSlugSanitizer();
-    $result = $sanitizer->sanitize('what--are------you-doing');
-
-    expect($result)->toBe('what-are-you-doing');
-});
-
-it('removes hyphens from the start and end', function () {
-    $sanitizer = new StringHelperSlugSanitizer();
-    $result = $sanitizer->sanitize('-what-are-you-doing-');
-
-    expect($result)->toBe('what-are-you-doing');
-});
